requirements.txt
Flask==2.3.3
pandas==2.1.1
openpyxl==3.1.2
xlrd==2.0.1
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7 